syscoffe/productos/api/serializers.py
from rest_framework import serializers
from ..models import Categoria, Producto

class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = ('id', 'nombre')